# docs/swagger3.0.yaml
openapi: 3.0.0
info:
  version: 1.0.16
  title: The Student Food forum
  description: >
    A forum API where accounts can post recipts about food and comment to them.  

    The application uses bearer token to authenticate requests.
    This token is created by requests to the /session router.

    All data is bound to the account and if ever deleted all other data will be
    cascaded.
    Comments are bound to the recipe and will also cascade if recipe is deleted.
  contact:
    name: "Filip Andersson, Marcus Gullstrand"
...
servers:
  - url: thesff.us-east-1.elasticbeanstalk.com
paths:
...
  /recipes:
    get:
      tags:
        - recipes
      summary: Get all recipes.
      description: ""
      parameters:
        - in: query
          name: recipesId
          description: Id of the recipes.
          schema:
            type: string
        - in: query
          name: title
          description: Title of the recipes.
          schema:
            type: string
        - in: query
          name: authorId
          description: Id of the author.
          schema:
            type: string
        - in: query
          name: limit
          description: How many recipes will be returned.
          schema:
            type: integer
            default: 25
        - in: query
          name: offset
          description: How many recipes that will be skipped.
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/recipes"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
    post:
      tags:
        - recipes
      summary: Create new recipes.
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
      description: ""
      responses:
        "201":
          description: Recipes created!
          headers:
            Location:
              schema:
                type: string
        "400":
          description: Bad request!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "404":
          description: Cannot find account!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                accountId:
                  description: The id of the account.
                  type: string
                title:
                  description: The title of the recipe.
                  type: string
                content:
                  description: The description of the recipe.
                  type: string
                image:
                  description: Add the recipe image. If left blank it will not be added.
                  type: string
                  format: binary
              required:
                - accountId
                - title
                - content
  "/recipes/{recipeId}":
    patch:
      tags:
        - recipes
      summary: Edit a given post.
      description: ""
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - name: recipeId
          in: path
          description: The id of the recipe.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully changed!
        "400":
          description: Bad request!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "404":
          description: Recipes not found!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  description: Edited title. If left blank it will not be edited.
                  type: string
                content:
                  description: Edited description. If left blank it will not be edited.
                  type: string
                image:
                  description: >-
                    Add/change the recipe image. If left blank it will not be
                    added/changed.
                  type: string
                  format: binary
    delete:
      tags:
        - recipes
      summary: Delete a given recipes.
      description: ""
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - name: recipeId
          in: path
          description: RecipeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully deleted recipes!
        "400":
          description: Bad request!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "404":
          description: Recipe not found!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
...
  /votes:
    get:
      tags:
        - votes
      summary: Get votes from one user.
      description: Gets all the users own votes
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
        - in: query
          name: accountId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          description: How many votes will be returned.
          schema:
            type: integer
            default: 25
        - in: query
          name: offset
          description: How many votes that will be skipped.
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/getVotes"
        "400":
          description: Missing parameter!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
    post:
      tags:
        - votes
      summary: Create a vote from an account.
      description: |
        "Used to create a vote from an account to a recipe.
         If a vote already exists, the preexisting vote will be updated with the new vote value."
      parameters:
        - in: header
          name: Authorization
          description: Bearer JWT Access Token
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Vote created!
        "400":
          description: Missing requestParameter!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "401":
          description: Unauthorized!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
        "500":
          description: Internal server error!
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/error"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postVotes"
        description: >-
          Vote paramiter takes -1, 0 or 1 as vaild values. -1 equals a downvote,
          0 equals a neutral vote, and 1 equals an upvote.
...
components:
  schemas:
    error:
      type: object
      properties:
        error:
          type: string
...
    recipes:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        imageId:
          type: string
          format: path
        updatedAt:
          type: string
      required:
        - accountId
...
    getVotes:
      type: object
      properties:
        title:
          type: string
        recipeId:
          type: string
        vote:
          type: integer
          format: smallint
    postVotes:
      type: object
      properties:
        accountId:
          type: string
        recipeId:
          type: string
        vote:
          type: integer
          format: smallint
          values: '"vote: 0", "vote: 1", "vote: -1"'
      required:
        - accountId
        - recipeId
        - vote
...